rename config file and fix some imports
